+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/JavaSourceProcessor.java
@@ -267,8 +267,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		if (!matcher.find()) {
 			processMessage(
 				fileName,
-				"LPS-49552: Missing override of BasePersistenceImpl." +
-					"fetchByPrimaryKeys(Set<Serializable>)");
+				"Missing override of BasePersistenceImpl." +
+					"fetchByPrimaryKeys(Set<Serializable>), see LPS-49552");
 		}
 	}
 
@@ -354,7 +354,10 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		if (!content.contains(
 				"package " + packagePath + StringPool.SEMICOLON)) {
 
-			processMessage(fileName, "Incorrect package path");
+			processMessage(
+				fileName,
+				"Package path does not match expected package path '" +
+					packagePath + "'");
 		}
 	}
 
@@ -425,8 +428,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			if (!localServiceImplContent.contains("@SystemEvent")) {
 				processMessage(
 					fileName,
-					"Missing deletion system event: " +
-						localServiceImplFileName);
+					"Missing deletion system event '" +
+						localServiceImplFileName + "', see LPS-46632");
 			}
 		}
 	}
@@ -455,7 +458,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 		if (pos != -1) {
 			processMessage(
-				fileName, "Use rs.getTimeStamp", getLineCount(content, pos));
+				fileName, "Use rs.getTimeStamp instead of rs.getDate",
+				getLineCount(content, pos));
 		}
 
 		// LPS-34911
@@ -470,7 +474,10 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			if (pos != -1) {
 				processMessage(
-					fileName, "ServiceUtil", getLineCount(content, pos));
+					fileName,
+					"Do not use *ServiceUtil classes in upgrade classes, see " +
+						"LPS-34911",
+					getLineCount(content, pos));
 			}
 		}
 
@@ -487,7 +494,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 				String componentAnnotation = matcher.group();
 
 				if (!componentAnnotation.contains("service =")) {
-					processMessage(fileName, "Missing service in @Component");
+					processMessage(
+						fileName, "@Component requires 'service' parameter");
 				}
 			}
 		}
@@ -523,8 +531,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 					processMessage(
 						fileName,
-						"LPS-65685: Break up Upgrade classes with a minor " +
-							"version increment or order alphabetically",
+						"Break up Upgrade classes with a minor version " +
+							"increment or order alphabetically, see LPS-65685",
 						getLineCount(content, matcher1.start()));
 
 					break;
@@ -729,7 +737,10 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			className.endsWith("ServiceImpl") &&
 			newContent.contains("ServiceUtil.")) {
 
-			processMessage(fileName, "ServiceUtil");
+			processMessage(
+				fileName,
+				"Do not use *ServiceUtil in *ServiceImpl class, create a " +
+					"reference via service.xml instead");
 		}
 
 		boolean isRunOutsidePortalExclusion = isExcludedPath(
@@ -739,15 +750,17 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			!isExcludedPath(_proxyExcludes, absolutePath) &&
 			newContent.contains("import java.lang.reflect.Proxy;")) {
 
-			processMessage(fileName, "Proxy");
+			processMessage(
+				fileName, "Use ProxyUtil instead of java.lang.reflect.Proxy");
 		}
 
 		if (newContent.contains("import edu.emory.mathcs.backport.java")) {
-			processMessage(fileName, "edu.emory.mathcs.backport.java");
+			processMessage(
+				fileName, "Illegal import: edu.emory.mathcs.backport.java");
 		}
 
 		if (newContent.contains("import jodd.util.StringPool")) {
-			processMessage(fileName, "jodd.util.StringPool");
+			processMessage(fileName, "Illegal import: jodd.util.StringPool");
 		}
 
 		// LPS-45027
@@ -785,7 +798,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			}
 
 			if ((pos3 < pos4) && (pos4 < pos5)) {
-				processMessage(fileName, "Use getInt(1) for count");
+				processMessage(
+					fileName, "Use rs.getInt(1) for count, see LPS-28266");
 			}
 		}
 
@@ -867,7 +881,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			processMessage(
 				fileName,
-				"Never import javax.servlet.jsp.* from portal-kernel");
+				"Never import javax.servlet.jsp.* from portal-kernel, see " +
+					"LPS-47682");
 		}
 
 		// LPS-48153
@@ -903,7 +918,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 		if (newContent.contains("org.testng.Assert")) {
 			processMessage(
-				fileName, "Use org.junit.Assert instead of org.testng.Assert");
+				fileName,
+				"Use org.junit.Assert instead of org.testng.Assert, see " +
+					"LPS-55690");
 		}
 
 		if (portalSource && isModulesFile(absolutePath) &&
@@ -925,7 +942,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			processMessage(
 				fileName,
 				"Use AutoBatchPreparedStatementUtil instead of " +
-					"DatabaseMetaData.supportsBatchUpdates");
+					"DatabaseMetaData.supportsBatchUpdates, see LPS-60473");
 		}
 
 		// LPS-64056
@@ -936,7 +953,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			processMessage(
 				fileName,
 				"Use ConfigurableUtil.createConfigurable instead of " +
-					"Configurable.createConfigurable");
+					"Configurable.createConfigurable, see LPS-64056");
 		}
 
 		// LPS-62786
@@ -1292,7 +1309,11 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			if (getLevel(matcher.group()) == 0) {
 				int lineCount = getLineCount(content, matcher.start());
 
-				processMessage(fileName, "line break", lineCount);
+				processMessage(
+					fileName,
+					"There should be line break before '" + matcher.group(1) +
+						"'",
+					lineCount);
 			}
 		}
 
@@ -1314,7 +1335,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			if (lastCharacterPreviousLine.equals(StringPool.OPEN_PARENTHESIS)) {
 				processMessage(
-					fileName, "line break",
+					fileName, "Line should not start with ')'",
 					getLineCount(content, matcher.start(1)));
 
 				return content;
@@ -1699,6 +1720,46 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		return content;
 	}
 
+	protected String formatDeprecatedJavadoc(String line) {
+		Matcher matcher = _deprecatedPattern.matcher(line);
+
+		if (!matcher.find()) {
+			return line;
+		}
+
+		ComparableVersion mainReleaseComparableVersion =
+			getMainReleaseComparableVersion();
+
+		if (matcher.group(2) == null) {
+			return StringUtil.insert(
+				line, " As of " + mainReleaseComparableVersion.toString(),
+				matcher.end(1));
+		}
+
+		String version = matcher.group(3);
+
+		ComparableVersion comparableVersion = new ComparableVersion(version);
+
+		if (comparableVersion.compareTo(mainReleaseComparableVersion) > 0) {
+			return StringUtil.replaceFirst(
+				line, version, mainReleaseComparableVersion.toString());
+		}
+
+		if (StringUtil.count(version, CharPool.PERIOD) == 1) {
+			return StringUtil.insert(line, ".0", matcher.end(3));
+		}
+
+		String deprecatedInfo = matcher.group(4);
+
+		if ((deprecatedInfo != null) &&
+			!deprecatedInfo.startsWith(StringPool.COMMA)) {
+
+			return StringUtil.insert(line, StringPool.COMMA, matcher.end(3));
+		}
+
+		return line;
+	}
+
 	protected String formatDuplicateReferenceMethods(
 			String fileName, String content, String className,
 			String packagePath)
@@ -1912,7 +1973,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 					int lineCount = getLineCount(content, matcher.start(2));
 
 					processMessage(
-						fileName, "Unprocessed " + originalExceptionClassName,
+						fileName,
+						"Unprocessed " + originalExceptionClassName +
+							", see LPS-36174",
 						lineCount);
 
 					break;
@@ -2215,7 +2278,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 				if (trimmedLine.contains("StopWatch stopWatch = null;")) {
 					processMessage(
-						fileName, "Do not set stopwatch to null", lineCount);
+						fileName, "Do not set stopwatch to null, see LPS-45492",
+						lineCount);
 				}
 
 				checkEmptyCollection(trimmedLine, fileName, lineCount);
@@ -2226,44 +2290,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 				checkResourceUtil(line, fileName, lineCount);
 
-				if (trimmedLine.startsWith("* @deprecated") &&
-					_addMissingDeprecationReleaseVersion) {
-
-					ComparableVersion mainReleaseComparableVersion =
-						getMainReleaseComparableVersion();
-
-					if (!trimmedLine.startsWith("* @deprecated As of ")) {
-						line = StringUtil.replace(
-							line, "* @deprecated",
-							"* @deprecated As of " +
-								mainReleaseComparableVersion.toString());
-					}
-					else {
-						String version = trimmedLine.substring(20);
-
-						version = StringUtil.split(
-							version, StringPool.SPACE)[0];
-
-						version = StringUtil.replace(
-							version, StringPool.COMMA, StringPool.BLANK);
-
-						ComparableVersion comparableVersion =
-							new ComparableVersion(version);
-
-						if (comparableVersion.compareTo(
-								mainReleaseComparableVersion) > 0) {
-
-							line = StringUtil.replaceFirst(
-								line, version,
-								mainReleaseComparableVersion.toString());
-						}
-						else if (StringUtil.count(
-									version, CharPool.PERIOD) == 1) {
-
-							line = StringUtil.replaceFirst(
-								line, version, version + ".0");
-						}
-					}
+				if (_addMissingDeprecationReleaseVersion) {
+					line = formatDeprecatedJavadoc(line);
 				}
 
 				if (trimmedLine.startsWith("* @see ") &&
@@ -2278,21 +2306,25 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 					line, fileName, absolutePath, lineCount, true);
 
 				if (trimmedLine.startsWith(StringPool.EQUAL)) {
-					processMessage(fileName, "line break", lineCount);
+					processMessage(
+						fileName, "Line should not start with '='", lineCount);
 				}
 
 				if (trimmedLine.startsWith("},") && !trimmedLine.equals("},")) {
-					processMessage(fileName, "line break", lineCount);
+					processMessage(
+						fileName, "There should be a line break after '},'",
+						lineCount);
 				}
 
 				if (line.contains("ActionForm form")) {
 					processMessage(
-						fileName, "Rename form to actionForm", lineCount);
+						fileName, "Rename 'form' to 'actionForm'", lineCount);
 				}
 
 				if (line.contains("ActionMapping mapping")) {
 					processMessage(
-						fileName, "Rename mapping to ActionMapping", lineCount);
+						fileName, "Rename 'mapping' to 'ActionMapping'",
+						lineCount);
 				}
 
 				if (!trimmedLine.equals("{") && line.endsWith("{") &&
@@ -2321,12 +2353,22 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 						(strippedQuotesLineOpenParenthesisCount > 0) &&
 						(getLevel(trimmedLine) > 0)) {
 
-						processMessage(fileName, "line break", lineCount);
+						processMessage(
+							fileName, "Incorrect line break", lineCount);
 					}
 
 					if (line.endsWith(StringPool.OPEN_PARENTHESIS)) {
-						if (line.contains(" && ") || line.contains(" || ")) {
-							processMessage(fileName, "line break", lineCount);
+						int x = line.lastIndexOf(" && ");
+						int y = line.lastIndexOf(" || ");
+
+						int z = Math.max(x, y);
+
+						if (z != -1) {
+							processMessage(
+								fileName,
+								"There should be a line break after '" +
+									line.substring(z + 1, z + 3) + "'",
+								lineCount);
 						}
 
 						int pos = strippedQuotesLine.indexOf(" + ");
@@ -2339,7 +2381,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 								(getLevel(linePart, "[", "]") == 0)) {
 
 								processMessage(
-									fileName, "line break", lineCount);
+									fileName,
+									"There should be a line break after '+'",
+									lineCount);
 							}
 						}
 					}
@@ -2428,15 +2472,20 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 						int x = -1;
 
 						while (true) {
-							x = strippedQuotesLine.indexOf(
+							x = trimmedLine.indexOf(
 								StringPool.COMMA_AND_SPACE, x + 1);
 
 							if (x == -1) {
 								break;
 							}
 
-							int level = getLevel(
-								strippedQuotesLine.substring(0, x));
+							if (ToolsUtil.isInsideQuotes(trimmedLine, x)) {
+								continue;
+							}
+
+							String linePart = trimmedLine.substring(0, x + 1);
+
+							int level = getLevel(linePart);
 
 							if ((previousLine.endsWith(
 									StringPool.OPEN_PARENTHESIS) &&
@@ -2445,25 +2494,35 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 								 (level <= 0))) {
 
 								processMessage(
-									fileName, "line break", lineCount);
+									fileName,
+									"There should be a line break after '" +
+										linePart + "'",
+									lineCount);
 							}
 						}
 					}
 
-					int x = strippedQuotesLine.indexOf(", ");
+					int x = trimmedLine.indexOf(StringPool.COMMA_AND_SPACE);
 
 					if (x != -1) {
-						String linePart = strippedQuotesLine.substring(0, x);
+						if (!ToolsUtil.isInsideQuotes(trimmedLine, x)) {
+							String linePart = trimmedLine.substring(0, x + 1);
 
-						if (getLevel(linePart) < 0) {
-							processMessage(fileName, "line break", lineCount);
+							if (getLevel(linePart) < 0) {
+								processMessage(
+									fileName,
+									"There should be a line break after '" +
+										linePart + "'",
+									lineCount);
+							}
 						}
 					}
 					else if (trimmedLine.endsWith(StringPool.COMMA) &&
 							 !trimmedLine.startsWith("for (")) {
 
 						if (getLevel(trimmedLine) > 0) {
-							processMessage(fileName, "line break", lineCount);
+							processMessage(
+								fileName, "Incorrect line break", lineCount);
 						}
 					}
 
@@ -2477,7 +2536,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 							if ((y == -1) || (x < y)) {
 								processMessage(
-									fileName, "line break", lineCount);
+									fileName,
+									"There should be a line break after '='",
+									lineCount);
 							}
 						}
 					}
@@ -2489,30 +2550,47 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 						 (line.endsWith(StringPool.OPEN_CURLY_BRACE) ||
 						  line.endsWith(StringPool.SEMICOLON)))) {
 
-						processMessage(fileName, "line break", lineCount);
+						processMessage(
+							fileName,
+							"There should be a line break before 'throws'",
+							lineCount);
 					}
 
-					if (trimmedLine.startsWith(StringPool.PERIOD) ||
-						(line.endsWith(StringPool.PERIOD) &&
-						 line.contains(StringPool.EQUAL))) {
+					if (trimmedLine.startsWith(StringPool.PERIOD)) {
+						processMessage(
+							fileName, "Line should not start with '.'",
+							lineCount);
+					}
 
-						processMessage(fileName, "line break", lineCount);
+					if (line.endsWith(StringPool.PERIOD) &&
+						line.contains(StringPool.EQUAL)) {
+
+						processMessage(
+							fileName, "There should be a line break after '='",
+							lineCount);
 					}
 
 					if (trimmedLine.matches("^\\} (catch|else|finally) .*")) {
-						processMessage(fileName, "line break", lineCount);
+						processMessage(
+							fileName, "There should be a line break after '{'",
+							lineCount);
 					}
 
 					if (previousLine.endsWith(StringPool.OPEN_PARENTHESIS) &&
 						trimmedLine.startsWith(StringPool.CLOSE_PARENTHESIS)) {
 
-						processMessage(fileName, "line break", lineCount);
+						processMessage(
+							fileName, "Line should not start with ')'",
+							lineCount);
 					}
 				}
 
-				if (line.contains("    ") && !line.matches("\\s*\\*.*")) {
+				if (line.contains(StringPool.FOUR_SPACES) &&
+					!line.matches("\\s*\\*.*")) {
+
 					if (!fileName.endsWith("StringPool.java")) {
-						processMessage(fileName, "tab", lineCount);
+						processMessage(
+							fileName, "Use tabs instead of spaces", lineCount);
 					}
 				}
 
@@ -2599,7 +2677,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 									previousLineLeadingTabCount)) {
 
 								processMessage(
-									fileName, "line break", lineCount);
+									fileName,
+									"There should be a line break after '('",
+									lineCount - 1);
 							}
 
 							if ((lineLeadingTabCount ==
@@ -2608,7 +2688,9 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 								 previousLine.endsWith(
 									 StringPool.OPEN_PARENTHESIS))) {
 
-								processMessage(fileName, "tab", lineCount);
+								processMessage(
+									fileName, "Line should be indented ",
+									lineCount);
 							}
 
 							if (Validator.isNotNull(trimmedLine) &&
@@ -2619,7 +2701,12 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 								((previousLineLeadingTabCount + 1) !=
 									lineLeadingTabCount)) {
 
-								processMessage(fileName, "tab", lineCount);
+								processMessage(
+									fileName,
+									"Line starts with " + lineLeadingTabCount +
+										" tabs, but should be " +
+											(previousLineLeadingTabCount + 1),
+									lineCount);
 							}
 
 							if (previousLine.endsWith(StringPool.PERIOD)) {
@@ -2635,18 +2722,20 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 										 CharPool.CLOSE_PARENTHESIS))) {
 
 									processMessage(
-										fileName, "line break", lineCount);
+										fileName, "Incorrect line break",
+										lineCount);
 								}
 							}
 
 							int diff =
 								lineLeadingTabCount -
-								previousLineLeadingTabCount;
+									previousLineLeadingTabCount;
 
 							if (trimmedLine.startsWith("throws ") &&
 								((diff == 0) || (diff > 1))) {
 
-								processMessage(fileName, "tab", lineCount);
+								processMessage(
+									fileName, "incorrect indent", lineCount);
 							}
 
 							if ((diff == 2) &&
@@ -2715,14 +2804,15 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 					processMessage(
 						fileName,
-						"Do not use PowerMock inside Arquillian tests");
+						"Do not use PowerMock inside Arquillian tests, see " +
+							"LPS-56706");
 				}
 
 				if (!packagePath.endsWith(".test")) {
 					processMessage(
 						fileName,
 						"Module integration test must be under a test " +
-							"subpackage");
+							"subpackage, see LPS-57722");
 				}
 			}
 			else if ((absolutePath.contains("/test/unit/") ||
@@ -2731,7 +2821,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 				processMessage(
 					fileName,
-					"Module unit test should not be under a test subpackage");
+					"Module unit test should not be under a test subpackage, " +
+						"see LPS-57722");
 			}
 		}
 
@@ -2740,7 +2831,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		if (content.contains("ProxyFactory.newServiceTrackedInstance(")) {
 			processMessage(
 				fileName,
-				"Do not use ProxyFactory.newServiceTrackedInstance in modules");
+				"Do not use ProxyFactory.newServiceTrackedInstance in " +
+					"modules, see LPS-57358");
 		}
 
 		// LPS-59076
@@ -2764,7 +2856,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			if (matcher.find()) {
 				processMessage(
 					fileName,
-					"Do not use com.liferay.registry classes in modules");
+					"Do not use com.liferay.registry classes in modules, " +
+						"see LPS-62989");
 			}
 		}
 
@@ -2775,7 +2868,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			processMessage(
 				fileName,
-				"Specify category using @ExtendedObjectClassDefinition");
+				"Specify category using @ExtendedObjectClassDefinition, see " +
+					"LPS-60186");
 		}
 
 		// LPS-64238
@@ -2783,14 +2877,17 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		if (content.contains("import com.liferay.util.dao.orm.CustomSQLUtil")) {
 			processMessage(
 				fileName,
-				"Do not use com.liferay.util.dao.orm.CustomSQLUtil in modules");
+				"Do not use com.liferay.util.dao.orm.CustomSQLUtil in " +
+					"modules, see LPS-64238");
 		}
 
 		// LPS-64335
 
 		if (content.contains("import com.liferay.util.ContentUtil")) {
 			processMessage(
-				fileName, "Do not use com.liferay.util.ContentUtil in modules");
+				fileName,
+				"Do not use com.liferay.util.ContentUtil in modules, see " +
+					"LPS-64335");
 		}
 
 		// LPS-67042
@@ -2829,8 +2926,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			processMessage(
 				fileName,
-				"LPS-59076: Use @Reference instead of calling " +
-					serviceUtilClassName + " directly");
+				"Use @Reference instead of calling " + serviceUtilClassName +
+					" directly, see LPS-59076");
 		}
 
 		matcher = _referenceMethodPattern.matcher(content);
@@ -3052,7 +3149,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 				}
 			}
 			else {
-				processMessage(fileName, "line break", lineCount);
+				processMessage(fileName, "Incorrect line break", lineCount);
 
 				return null;
 			}
@@ -3187,7 +3284,11 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 			if (previousLine.endsWith(StringPool.EQUAL)) {
 				if (line.endsWith(StringPool.OPEN_CURLY_BRACE)) {
-					processMessage(fileName, "line break", lineCount);
+					processMessage(
+						fileName,
+						"'" + trimmedLine + "' should be added to previous " +
+							"line",
+						lineCount);
 
 					return null;
 				}
@@ -3201,7 +3302,11 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 						if (Validator.isNull(nextLine) ||
 							nextLine.endsWith(") {")) {
 
-							processMessage(fileName, "line break", lineCount);
+							processMessage(
+								fileName,
+								"'" + trimmedLine + "' should be added to " +
+									"previous line",
+								lineCount);
 
 							return null;
 						}
@@ -3236,10 +3341,10 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		}
 
 		if (((trimmedLine.length() + previousLineLength) <= _maxLineLength) &&
-			(previousLine.endsWith(StringPool.OPEN_BRACKET) ||
-			 previousLine.endsWith(StringPool.OPEN_PARENTHESIS) ||
-			 previousLine.endsWith(StringPool.PERIOD)) &&
-			line.endsWith(StringPool.SEMICOLON)) {
+			(previousLine.endsWith(StringPool.PERIOD) ||
+			 (previousLine.endsWith(StringPool.OPEN_BRACKET) ||
+			  previousLine.endsWith(StringPool.OPEN_PARENTHESIS)) &&
+			 line.endsWith(StringPool.SEMICOLON))) {
 
 			return getCombinedLinesContent(
 				content, fileName, line, trimmedLine, lineLength, lineCount,
@@ -3352,7 +3457,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 				}
 
 				if (trimmedLine.equals(linePart)) {
-					processMessage(fileName, "line break", lineCount);
+					processMessage(fileName, "Incorrect line break", lineCount);
 
 					return null;
 				}
@@ -4366,7 +4471,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 		File baseDirFile = new File(sourceFormatterArgs.getBaseDirName());
 
-		List<SourceFormatterMessage> sourceFormatterMessages =
+		Set<SourceFormatterMessage> sourceFormatterMessages =
 			CheckStyleUtil.process(
 				_ungeneratedFiles, getAbsolutePath(baseDirFile));
 
@@ -4490,6 +4595,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		"@Component(\n|\\([\\s\\S]*?\\)\n)");
 	private final Pattern _customSQLFilePattern = Pattern.compile(
 		"<sql file=\"(.*)\" \\/>");
+	private final Pattern _deprecatedPattern = Pattern.compile(
+		"(^\\s*\\* @deprecated)( As of ([0-9\\.]+)(.+)?)?");
 	private List<String> _diamondOperatorExcludes;
 	private final Pattern _diamondOperatorPattern = Pattern.compile(
 		"(return|=)\n?(\t+| )new ([A-Za-z]+)(\\s*)<(.+)>\\(\n*\t*.*\\);\n");
@@ -4508,7 +4615,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 	private final Pattern _incorrectLineBreakPattern2 = Pattern.compile(
 		"\\{\n\n\t*\\}");
 	private final Pattern _incorrectLineBreakPattern3 = Pattern.compile(
-		", new .*\\(.*\\) \\{\n");
+		", (new .*\\(.*\\) \\{)\n");
 	private final Pattern _incorrectLineBreakPattern4 = Pattern.compile(
 		"\n(\t*)(.*\\) \\{)([\t ]*\\}\n)");
 	private final Pattern _incorrectLineBreakPattern5 = Pattern.compile(